refactor: change check (only check is jalon IS NOT date_personnalisee, since all other jalons are non editable)
